% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basemap}
\alias{basemap}
\title{mapa base de Catalunya (OpenStreetMap)}
\format{
## `basemap`
stars object with 3 dimensions and 1 attribute:
}
\source{
"Openstreetmap Collaborators"
}
\usage{
basemap
}
\description{
Objecte stars utilitzat com a mapa base. Font: Col·laboradors d'OpenStreetMap
}
\keyword{datasets}
