% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat}
\alias{cat}
\title{límits de Catalunya (Idescat)}
\format{
## `cat`
Simple feature collection with 1 feature and 0 fields
 Geometry type: MULTIPOLYGON
 Dimension:     XY
 Bounding box:  xmin: 0.1594159 ymin: 40.52289 xmax: 3.332551 ymax: 42.86149
 Geodetic CRS:  WGS 84
}
\source{
"Idescat"
}
\usage{
cat
}
\description{
Objecte sf que recull un polígon amb tota la superfície de Catalunya
}
\keyword{datasets}
