% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corrplot.R
\name{make_corrplot}
\alias{make_corrplot}
\title{make_corrplot}
\usage{
make_corrplot(x, y)
}
\arguments{
\item{x}{a character vector with all independent variables}

\item{y}{y a character vector with the dependent variable of interest}
}
\value{
a correlation plot
}
\description{
makes a correlation plot from selected variables
}
\examples{
\dontrun{
make_corrplot(c("hotels_1000hab", "vab_relatiu_Agricultura", "renda_capita"), "consum_total")
}
}
