% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{make_plots}
\alias{make_plots}
\title{make_plots}
\usage{
make_plots(var1, var2)
}
\arguments{
\item{var1}{variable 1 que es vol estudiar}

\item{var2}{variable 2 que es vol estudiar}
}
\value{
a plot grid with two maps and a scatterplot
}
\description{
retorna dos mapes coropleta un per cada variable i el gràfic de dispersió
}
