% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{munis_consum}
\alias{munis_consum}
\title{Dades de consum d'aigua i variables explicatives}
\format{
## `munis_consum`
Simple feature collection with 532 features and 38 fields
 Geometry type: MULTIPOLYGON
 Dimension:     XY
 Bounding box:  xmin: 0.6423511 ymin: 40.76635 xmax: 3.332551 ymax: 42.46151
 Geodetic CRS:  WGS 84
}
\source{
"Idescat, INE, OpenStreetMap, Generalitat"
}
\usage{
munis_consum
}
\description{
Objecte sf que recull les dades de municipis de conques internes amb dades de consum d'aigua
}
\keyword{datasets}
