% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{make_plots}
\alias{make_plots}
\title{make_plots}
\usage{
make_plots(x, y)
}
\arguments{
\item{x}{a character vector with all independent variables}

\item{y}{y a character vector with the dependent variable of interest}
}
\value{
a faceted choropleth map
}
\description{
makes choropleth maps with all variables of interest
}
